cleanup path utils
