+++ python/ray/tests/test_runtime_env_conda_and_pip.py
@@ -9,13 +9,14 @@ from ray._private.test_utils import (
     check_local_files_gced,
     generate_runtime_env_dict,
 )
+from ray._private.runtime_env import dependency_utils
 from ray._private.runtime_env.conda import _get_conda_dict_with_ray_inserted
-from ray._private.runtime_env.pip import (
+from ray._private.runtime_env.dependency_utils import (
     INTERNAL_PIP_FILENAME,
     MAX_INTERNAL_PIP_FILENAME_TRIES,
-    _PathHelper,
 )
 from ray.runtime_env import RuntimeEnv
+from ray._private.runtime_env import virtualenv_utils
 
 import yaml
 import tempfile
@@ -231,27 +232,25 @@ def test_runtime_env_conda_not_exists_not_hang(shutdown_only):
 
 
 def test_get_requirements_file():
-    """Unit test for _PathHelper.get_requirements_file."""
+    """Unit test for virtualenv_utils.get_requirements_file."""
     with tempfile.TemporaryDirectory() as tmpdir:
-        path_helper = _PathHelper()
-
         # If pip_list is None, we should return the internal pip filename.
-        assert path_helper.get_requirements_file(tmpdir, pip_list=None) == os.path.join(
-            tmpdir, INTERNAL_PIP_FILENAME
-        )
+        assert dependency_utils.get_requirements_file(
+            tmpdir, pip_list=None
+        ) == os.path.join(tmpdir, INTERNAL_PIP_FILENAME)
 
         # If the internal pip filename is not in pip_list, we should return the internal
         # pip filename.
-        assert path_helper.get_requirements_file(
+        assert dependency_utils.get_requirements_file(
             tmpdir, pip_list=["foo", "bar"]
         ) == os.path.join(tmpdir, INTERNAL_PIP_FILENAME)
 
         # If the internal pip filename is in pip_list, we should append numbers to the
         # end of the filename until we find one that doesn't conflict.
-        assert path_helper.get_requirements_file(
+        assert dependency_utils.get_requirements_file(
             tmpdir, pip_list=["foo", "bar", f"-r {INTERNAL_PIP_FILENAME}"]
         ) == os.path.join(tmpdir, f"{INTERNAL_PIP_FILENAME}.1")
-        assert path_helper.get_requirements_file(
+        assert virtualenv_utils.get_requirements_file(
             tmpdir,
             pip_list=[
                 "foo",
@@ -263,7 +262,7 @@ def test_get_requirements_file():
 
         # If we can't find a valid filename, we should raise an error.
         with pytest.raises(RuntimeError) as excinfo:
-            path_helper.get_requirements_file(
+            dependency_utils.get_requirements_file(
                 tmpdir,
                 pip_list=[
                     "foo",
